telegram_bot/requirements.txt
python-telegram-bot[job-queue]==21.1.1
requests
pandas
matplotlib